Defer installation of doct hook functions
Previous approach had doct declaring org-capture-*-hook and setting it
to nil so we could add to it when loading doct without requiring
org-capture. It worked, but it then looks like doct 'owns' those
variables to the help system. This fix commits the sin of using
with-eval-after load to install them once org-capture is loaded. It also
adds the proper symbols instead of the function values to the list, so
they are readable from the help system. If anyone reading this knows a
better way, I'm all ears.