Ensure properties are strict plist of form (KEYWORD VAL...)
User was confused by doct's behavior when programatically setting a declaration's
:children. Reproduction of user's use-case:

(let* ((blog-topics '(("Lifestyle" "l")
                      ("Tech" "t")
                      ("Fitness" "f")))
       (org-capture-templates
        (doct `(:group
                :empty-lines 1
                :children
                (("Blog"
                  :keys "b"
                  :file ""
                  :type entry
                  :template "* TODO %^{name}\n  %i%?\n  %a"
                  :children ,@(cl-loop
                               for (key value)
                               in (sort blog-topics (lambda (a b) (string< (car a) (car b))))
                               collect (list key :keys value :headline key)))))))))

The list of children was spliced in a backquoted form, rather
than inserted as a whole list. Since this was technically a valid plist,
doct silently ignored the extra children and used only the first,
associated child. This was not (and should not be) the intended
behavior. The properties are now checked with doct--plist-p, which
ensures strict (KEYWORD VAL..) form.