Fix :doct-custom inheritance
If a parent had a :custom plist, the plist's properties would override the
child's properties unless they were also declared in a :custom plist. e.g.

("parent" :custom (:property t)
          :children ("child" :property nil))

Would incorrectly return t when filling the "child" template. This is because we
were appending to the :doct-custom plist in the case of inheritance, which would
allow for the same key to appear multiple times. `plist-get' will return the
first found key's value (in this case the "parent" declaration's).

See: #28