doct--target-file: Get :olp path when :datetree declared first
If :datetree is declared before :olp, we need to explicitly fetch the :olp value.
Otherwise, the :olp value was silently dropped from the declaration.
This didn't raise any errors because nil is a valid :olp value (in order to
override any :olp inheritance), and (file+olp+datetree "file/without/olp") is a
valid target.

See: #31